+++ np.CallbackURLs/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.CallbackURLs/README.md) for details on available commands and use cases.
 
+## [1.0.0] - 2022-07-11 (@dwertheimer)
+- Changed plugin Name to: "🧩 External Links, X-Callback-URLs, RunPlugin Creator"
+- Added command "Create Link to Current Note+Heading" with direct access from command bar
+
 ## [0.6.0] = 2022-07-02 (@dwertheimer)
 - Added noteInfo command
 - Added deleteNote command